requirements.txt
# necessary modules/packages/libraries for py file
pandas==1.5.*
mediacloud_metadata==0.9.*
wayback_news_search==1.0.*
requests
